CRITICAL FIX: Resolve API route 'window is not defined' errors
- Create shared types/api.ts to avoid circular imports between API routes and client modules
- Move RankingsResponse, DailyRankingsQuery, HourlyRankingsQuery types to shared file
- Update all imports to use shared types instead of importing from API route files
- Prevents client-side code from being pulled into server-side API context
- Fixes Node.js process crashes with ReferenceError: window is not defined

Root cause: modules/topCryptos.ts and modules/processRankings.ts were importing types
from pages/api/rankings/* files, creating circular dependencies that caused client-side
code to execute in server context during build/runtime.
